+++ airbyte-webapp/src/views/Connector/ConnectorForm/components/Controls/ConnectorServiceTypeControl/ConnectorDefinitionTypeControl.tsx
@@ -24,6 +24,7 @@ import { useAvailableConnectorDefinitions } from "hooks/domain/connector/useAvai
 import { useExperiment } from "hooks/services/Experiment";
 import { useModalService } from "hooks/services/Modal";
 import { useCurrentWorkspace } from "hooks/services/useWorkspace";
+import { InlineEnrollmentCallout } from "packages/cloud/components/experiments/FreeConnectorProgram/InlineEnrollmentCallout";
 import { useDocumentationPanelContext } from "views/Connector/ConnectorDocumentationLayout/DocumentationPanelContext";
 import RequestConnectorModal from "views/Connector/RequestConnectorModal";
 
@@ -202,7 +203,10 @@ export const ConnectorDefinitionTypeControl: React.FC<ConnectorDefinitionTypeCon
       {selectedConnectorDefinition &&
         (selectedConnectorDefinition.releaseStage === ReleaseStage.alpha ||
           selectedConnectorDefinition.releaseStage === ReleaseStage.beta) && (
-          <WarningMessage stage={selectedConnectorDefinition.releaseStage} />
+          <>
+            <InlineEnrollmentCallout />
+            <WarningMessage stage={selectedConnectorDefinition.releaseStage} />
+          </>
         )}
     </>
   );
